Split signedData into separate types
