+++ consensus/ibft/src/main/java/org/hyperledger/besu/consensus/ibft/messagewrappers/Prepare.java
@@ -14,6 +14,7 @@
  */
 package org.hyperledger.besu.consensus.ibft.messagewrappers;
 
+import org.hyperledger.besu.consensus.ibft.payload.PayloadDeserialisers;
 import org.hyperledger.besu.consensus.ibft.payload.PreparePayload;
 import org.hyperledger.besu.consensus.ibft.payload.SignedData;
 import org.hyperledger.besu.ethereum.core.Hash;
@@ -32,6 +33,6 @@ public class Prepare extends IbftMessage<PreparePayload> {
   }
 
   public static Prepare decode(final Bytes data) {
-    return new Prepare(SignedData.readSignedPreparePayloadFrom(RLP.input(data)));
+    return new Prepare(PayloadDeserialisers.readSignedPreparePayloadFrom(RLP.input(data)));
   }
 }
