+++ consensus/ibft/src/main/java/org/hyperledger/besu/consensus/ibft/payload/PayloadDeserialisers.java
@@ -1,0 +1,82 @@
+/*
+ * Copyright 2020 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package org.hyperledger.besu.consensus.ibft.payload;
+
+import org.hyperledger.besu.crypto.SECP256K1.Signature;
+import org.hyperledger.besu.ethereum.core.Address;
+import org.hyperledger.besu.ethereum.core.Util;
+import org.hyperledger.besu.ethereum.rlp.RLPInput;
+
+public class PayloadDeserialisers {
+
+  public static SignedData<ProposalPayload> readSignedProposalPayloadFrom(final RLPInput rlpInput) {
+
+    rlpInput.enterList();
+    final ProposalPayload unsignedMessageData = ProposalPayload.readFrom(rlpInput);
+    final Signature signature = readSignature(rlpInput);
+    rlpInput.leaveList();
+
+    return from(unsignedMessageData, signature);
+  }
+
+  public static SignedData<PreparePayload> readSignedPreparePayloadFrom(final RLPInput rlpInput) {
+
+    rlpInput.enterList();
+    final PreparePayload unsignedMessageData = PreparePayload.readFrom(rlpInput);
+    final Signature signature = readSignature(rlpInput);
+    rlpInput.leaveList();
+
+    return from(unsignedMessageData, signature);
+  }
+
+  public static SignedData<CommitPayload> readSignedCommitPayloadFrom(final RLPInput rlpInput) {
+
+    rlpInput.enterList();
+    final CommitPayload unsignedMessageData = CommitPayload.readFrom(rlpInput);
+    final Signature signature = readSignature(rlpInput);
+    rlpInput.leaveList();
+
+    return from(unsignedMessageData, signature);
+  }
+
+  public static SignedData<RoundChangePayload> readSignedRoundChangePayloadFrom(
+      final RLPInput rlpInput) {
+
+    rlpInput.enterList();
+    final RoundChangePayload unsignedMessageData = RoundChangePayload.readFrom(rlpInput);
+    final Signature signature = readSignature(rlpInput);
+    rlpInput.leaveList();
+
+    return from(unsignedMessageData, signature);
+  }
+
+  protected static <M extends Payload> SignedData<M> from(
+      final M unsignedMessageData, final Signature signature) {
+
+    final Address sender = recoverSender(unsignedMessageData, signature);
+
+    return new SignedData<>(unsignedMessageData, sender, signature);
+  }
+
+  protected static Signature readSignature(final RLPInput signedMessage) {
+    return signedMessage.readBytes(Signature::decode);
+  }
+
+  protected static Address recoverSender(
+      final Payload unsignedMessageData, final Signature signature) {
+
+    return Util.signatureToAddress(signature, MessageFactory.hashForSignature(unsignedMessageData));
+  }
+}
