+++ consensus/ibft/src/main/java/org/hyperledger/besu/consensus/ibft/payload/RoundChangeCertificate.java
@@ -38,7 +38,7 @@ public class RoundChangeCertificate {
     final List<SignedData<RoundChangePayload>> roundChangePayloads;
 
     rlpInput.enterList();
-    roundChangePayloads = rlpInput.readList(SignedData::readSignedRoundChangePayloadFrom);
+    roundChangePayloads = rlpInput.readList(PayloadDeserialisers::readSignedRoundChangePayloadFrom);
     rlpInput.leaveList();
 
     return new RoundChangeCertificate(roundChangePayloads);
