+++ consensus/ibft/src/test/java/org/hyperledger/besu/consensus/ibft/payload/RoundChangeCertificateTest.java
@@ -59,11 +59,13 @@ public class RoundChangeCertificateTest {
         TestHelpers.createProposalBlock(singletonList(AddressHelpers.ofValue(1)), ROUND_IDENTIFIER);
     final ProposalPayload proposalPayload = new ProposalPayload(ROUND_IDENTIFIER, block.getHash());
     final Signature signature = Signature.create(BigInteger.ONE, BigInteger.TEN, (byte) 0);
-    SignedData<ProposalPayload> signedProposal = SignedData.from(proposalPayload, signature);
+    SignedData<ProposalPayload> signedProposal =
+        PayloadDeserialisers.from(proposalPayload, signature);
 
     final PreparePayload preparePayload =
         new PreparePayload(ROUND_IDENTIFIER, Hash.fromHexStringLenient("0x8523ba6e7c5f59ae87"));
-    final SignedData<PreparePayload> signedPrepare = SignedData.from(preparePayload, signature);
+    final SignedData<PreparePayload> signedPrepare =
+        PayloadDeserialisers.from(preparePayload, signature);
 
     final PreparedCertificate preparedCert =
         new PreparedCertificate(signedProposal, Lists.newArrayList(signedPrepare));
